apps/gateway_admin/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "gateway-admin"
version = "0.0.1"
description = "Administrative CLI for the Geldstrom gateway"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gateway-contracts",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "argon2-cffi>=23.1",
    "typer>=0.15",
    "rich>=13.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "aiosmtplib>=3.0",
    "email-validator>=2.0",
]

[project.scripts]
gw-admin = "gateway_admin.presentation.cli.main:main"
gw-admin-api = "gateway_admin.presentation.api.main:app"

[tool.hatch.build.targets.wheel]
packages = ["gateway_admin"]

[tool.hatch.build.targets.sdist]
include = [
    "/gateway_admin",
    "/README.md",
    "/pyproject.toml",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21,<0.24",
    "httpx>=0.27",
    "hypothesis>=6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
